Removing useless changes
