+++ src/spikeinterface/sorters/internal/spyking_circus2.py
@@ -18,7 +18,10 @@ from spikeinterface.core.basesorting import minimum_spike_dtype
 from spikeinterface.core.sparsity import compute_sparsity
 from spikeinterface.sortingcomponents.tools import remove_empty_templates
 from spikeinterface.sortingcomponents.tools import get_prototype_spike, check_probe_for_drift_correction
-from spikeinterface.sortingcomponents.clustering.clustering_tools import final_cleaning_circus
+from spikeinterface.core.sortinganalyzer import create_sorting_analyzer
+from spikeinterface.curation.auto_merge import get_potential_auto_merge
+from spikeinterface.sortingcomponents.clustering.clustering_tools import resolve_merging_graph, apply_merges_to_sorting
+from spikeinterface.core.analyzer_extension_core import ComputeTemplates
 
 try:
     import hdbscan
@@ -331,3 +334,25 @@ class Spykingcircus2Sorter(ComponentsBasedSorter):
         sorting = sorting.save(folder=sorting_folder)
 
         return sorting
+
+
+def final_cleaning_circus(recording, sorting, templates, 
+                          **merging_kwargs):
+
+    
+
+    sparsity = templates.sparsity
+    templates_array = templates.get_dense_templates().copy()
+
+    sa = create_sorting_analyzer(sorting, recording, format="memory", sparsity=sparsity)
+    
+    sa.extensions['templates'] = ComputeTemplates(sa)
+    sa.extensions['templates'].params = {'nbefore' : templates.nbefore}
+    sa.extensions['templates'].data['average'] = templates_array
+    sa.compute('unit_locations', method='monopolar_triangulation')
+    merges = get_potential_auto_merge(sa, **merging_kwargs)
+    merges = resolve_merging_graph(sorting, merges)
+    sorting = apply_merges_to_sorting(sorting, merges)
+    #sorting = merge_units_sorting(sorting, merges)
+
+    return sorting
